+++ components/Header.tsx
@@ -17,7 +17,6 @@ import { TRANSITIONS } from 'theme'
 import { Box, Button, Card, Container, Flex, Grid, Image, SxStyleProp, Text } from 'theme-ui'
 
 import { ContextConnected } from '../blockchain/network'
-import { useFeatureToggle } from '../helpers/useFeatureToggle'
 import { useAppContext } from './AppContextProvider'
 import { ChevronUpDown } from './ChevronUpDown'
 import { useSharedUI } from './SharedUIProvider'
@@ -132,18 +131,15 @@ function navLinkColor(isActive: boolean) {
 
 function ConnectedHeader() {
   const { pathname } = useRouter()
-  const { vaultFormToggleTitle, setVaultFormOpened } = useSharedUI()
   const { accountData$, context$ } = useAppContext()
   const { t } = useTranslation()
   const accountData = useObservable(accountData$)
   const context = useObservable(context$)
-  const assetLandingPagesFeatureEnabled = useFeatureToggle('AssetLandingPages')
 
   const numberOfVaults =
     accountData?.numberOfVaults !== undefined ? accountData.numberOfVaults : undefined
-  const firstCDP = numberOfVaults ? numberOfVaults === 0 : undefined
 
-  return assetLandingPagesFeatureEnabled ? (
+  return (
     <>
       <Box sx={{ display: ['none', 'block'], mb: 5 }}>
         <BasicHeader
@@ -214,72 +210,6 @@ function ConnectedHeader() {
         </BasicHeader>
       </Box>
     </>
-  ) : (
-    <BasicHeader
-      sx={{
-        position: 'relative',
-        alignItems: 'center',
-        zIndex: 1,
-      }}
-      variant="appContainer"
-    >
-      <>
-        <Flex
-          sx={{
-            alignItems: 'center',
-            justifyContent: ['space-between', 'flex-start'],
-            width: ['100%', 'auto'],
-          }}
-        >
-          <Logo />
-          <Flex sx={{ ml: 5, zIndex: 1 }}>
-            <AppLink
-              variant="nav"
-              sx={{ mr: 4 }}
-              href={`/owner/${(context as ContextConnected)?.account}`}
-              onClick={() => trackingEvents.yourVaults()}
-            >
-              {t('your-vaults')} {numberOfVaults ? numberOfVaults > 0 && `(${numberOfVaults})` : ''}
-            </AppLink>
-            <AppLink
-              variant="nav"
-              sx={{ mr: [0, 4] }}
-              href="/vaults/list"
-              onClick={() => trackingEvents.createNewVault(firstCDP)}
-            >
-              <Box sx={{ display: ['none', 'block'] }}>{t('open-new-vault')}</Box>
-              <Box sx={{ display: ['block', 'none'], fontSize: '0px' }}>
-                <Icon name="plus_header" size="auto" width="18px" height="18px" />
-              </Box>
-            </AppLink>
-          </Flex>
-        </Flex>
-        <Flex
-          sx={{
-            position: ['fixed', 'relative'],
-            bottom: 0,
-            left: 0,
-            right: 0,
-            bg: ['rgba(255,255,255,0.9)', 'transparent'],
-            p: [3, 0],
-            justifyContent: 'space-between',
-            gap: 2,
-          }}
-        >
-          <Flex>
-            <UserSettingsButton />
-            <AccountButton />
-          </Flex>
-          {vaultFormToggleTitle && (
-            <Box sx={{ display: ['flex', 'none'] }}>
-              <Button variant="menuButton" sx={{ px: 3 }} onClick={() => setVaultFormOpened(true)}>
-                <Box>{vaultFormToggleTitle}</Box>
-              </Button>
-            </Box>
-          )}
-        </Flex>
-      </>
-    </BasicHeader>
   )
 }
 
@@ -424,23 +354,6 @@ const LangSelectMobileComponents: Partial<SelectComponents<{
   ),
 }
 
-const MOBILE_MENU_SECTIONS_PRE_ASSET_LANDING_PAGES = [
-  {
-    titleKey: 'nav.products',
-    links: [
-      { labelKey: 'nav.dai-wallet', url: HEADER_LINKS['dai-wallet'] },
-      { labelKey: 'nav.oasis-borrow' },
-    ],
-  },
-  {
-    titleKey: 'nav.resources',
-    links: [
-      { labelKey: 'nav.learn', url: HEADER_LINKS['learn'] },
-      { labelKey: 'nav.blog', url: HEADER_LINKS['blog'] },
-    ],
-  },
-]
-
 const MOBILE_MENU_DISCONNECTED_SECTIONS = [
   {
     titleKey: 'nav.products',
@@ -472,16 +385,11 @@ const MOBILE_MENU_CONNECTED_SECTIONS = [
 function MobileMenu() {
   const { t } = useTranslation()
   const { pathname } = useRouter()
-  const assetLandingPagesFeatureEnabled = useFeatureToggle('AssetLandingPages')
   const [isOpen, setIsOpen] = useState(false)
   const { context$ } = useAppContext()
   const context = useObservable(context$)
 
   const isConnected = !!(context as ContextConnected)?.account
-
-  const mobileMenuEntries = assetLandingPagesFeatureEnabled
-    ? MOBILE_MENU_DISCONNECTED_SECTIONS
-    : MOBILE_MENU_SECTIONS_PRE_ASSET_LANDING_PAGES
 
   const closeMenu = useCallback(() => setIsOpen(false), [])
   return (
@@ -516,7 +424,7 @@ function MobileMenu() {
       >
         <Grid sx={{ rowGap: 5, mt: 3, mx: 'auto', maxWidth: 7 }}>
           {!isConnected &&
-            mobileMenuEntries.map((section) => (
+            MOBILE_MENU_DISCONNECTED_SECTIONS.map((section) => (
               <Grid key={section.titleKey}>
                 <Text variant="links.navHeader">{t(section.titleKey)}</Text>
                 {section.links.map((link) =>
@@ -546,7 +454,6 @@ function MobileMenu() {
               </Grid>
             ))}
           {isConnected &&
-            assetLandingPagesFeatureEnabled &&
             MOBILE_MENU_CONNECTED_SECTIONS.map((section) => (
               <Grid key={section.titleKey}>
                 <Text variant="links.navHeader">{t(section.titleKey)}</Text>
@@ -595,61 +502,42 @@ function MobileMenu() {
 function DisconnectedHeader() {
   const { t } = useTranslation()
   const { pathname } = useRouter()
-  const assetLandingPagesEnabled = useFeatureToggle('AssetLandingPages')
-  const menuBarLandingPagesDisabled = (
-    <>
-      <HeaderDropdown title={t('nav.products')}>
-        <AppLink variant="links.nav" sx={{ fontWeight: 'body' }} href={HEADER_LINKS['dai-wallet']}>
-          {t('nav.dai-wallet')}
-        </AppLink>
-        <Text variant="links.nav" sx={{ cursor: 'default', ':hover': { color: 'primary' } }}>
-          {t('nav.borrow')}
-        </Text>
-      </HeaderDropdown>
-      <AppLink variant="links.navHeader" href={HEADER_LINKS['learn']}>
-        {t('nav.learn')}
-      </AppLink>
-      <AppLink variant="links.navHeader" href={HEADER_LINKS['blog']}>
-        {t('nav.blog')}
-      </AppLink>
-    </>
-  )
-  const menuBarLandingPagesEnabled = (
-    <>
-      <AppLink
-        variant="links.navHeader"
-        href={HEADER_LINKS.multiply}
-        sx={{ color: navLinkColor(pathname.includes(HEADER_LINKS.multiply)) }}
-      >
-        {t('nav.multiply')}
-      </AppLink>
-      <AppLink
-        variant="links.navHeader"
-        href={HEADER_LINKS.borrow}
-        sx={{ color: navLinkColor(pathname.includes(HEADER_LINKS.borrow)) }}
-      >
-        {t('nav.borrow')}
-      </AppLink>
-      <HeaderDropdown title={t('nav.more')}>
-        <AppLink variant="links.nav" sx={{ fontWeight: 'body' }} href={HEADER_LINKS['dai-wallet']}>
-          {t('nav.dai-wallet')}
-        </AppLink>
-        <AppLink variant="links.nav" sx={{ fontWeight: 'body' }} href={HEADER_LINKS['learn']}>
-          {t('nav.learn')}
-        </AppLink>
-        <AppLink variant="links.nav" sx={{ fontWeight: 'body' }} href={HEADER_LINKS['blog']}>
-          {t('nav.blog')}
-        </AppLink>
-      </HeaderDropdown>
-    </>
-  )
+
   return (
     <>
       <Box sx={{ display: ['none', 'block'] }}>
         <BasicHeader variant="appContainer">
           <Grid sx={{ alignItems: 'center', columnGap: [4, 4, 5], gridAutoFlow: 'column', mr: 3 }}>
             <Logo />
-            {assetLandingPagesEnabled ? menuBarLandingPagesEnabled : menuBarLandingPagesDisabled}
+            <AppLink
+              variant="links.navHeader"
+              href={HEADER_LINKS.multiply}
+              sx={{ color: navLinkColor(pathname.includes(HEADER_LINKS.multiply)) }}
+            >
+              {t('nav.multiply')}
+            </AppLink>
+            <AppLink
+              variant="links.navHeader"
+              href={HEADER_LINKS.borrow}
+              sx={{ color: navLinkColor(pathname.includes(HEADER_LINKS.borrow)) }}
+            >
+              {t('nav.borrow')}
+            </AppLink>
+            <HeaderDropdown title={t('nav.more')}>
+              <AppLink
+                variant="links.nav"
+                sx={{ fontWeight: 'body' }}
+                href={HEADER_LINKS['dai-wallet']}
+              >
+                {t('nav.dai-wallet')}
+              </AppLink>
+              <AppLink variant="links.nav" sx={{ fontWeight: 'body' }} href={HEADER_LINKS['learn']}>
+                {t('nav.learn')}
+              </AppLink>
+              <AppLink variant="links.nav" sx={{ fontWeight: 'body' }} href={HEADER_LINKS['blog']}>
+                {t('nav.blog')}
+              </AppLink>
+            </HeaderDropdown>
           </Grid>
           <Grid sx={{ alignItems: 'center', columnGap: 3, gridAutoFlow: 'column' }}>
             <AppLink
