+++ helpers/useFeatureToggle.ts
@@ -3,13 +3,11 @@ export const FT_LOCAL_STORAGE_KEY = 'features'
 
 type ConfiguredFeatures = Record<Features, boolean>
 
-type Features = 'TestFeature' | 'AnotherTestFeature' | 'AssetLandingPages' | 'EarnProduct'
+type Features = 'TestFeature' | 'AnotherTestFeature' | 'EarnProduct'
 const configuredFeatures: Record<Features, boolean> = {
   TestFeature: false, // used in unit tests
   AnotherTestFeature: true, // used in unit tests
-
   EarnProduct: false,
-  AssetLandingPages: true,
   // your feature here....
 }
