+++ pages/index.tsx
@@ -1,11 +1,9 @@
 import { WithConnection } from 'components/connectWallet/ConnectWallet'
 import { LandingPageLayout } from 'components/Layouts'
-import { LandingView } from 'features/landing/LandingView'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 import React from 'react'
 
 import { HomepageView } from '../features/homepage/HomepageView'
-import { useFeatureToggle } from '../helpers/useFeatureToggle'
 
 export const getStaticProps = async ({ locale }: { locale: string }) => ({
   props: {
@@ -14,9 +12,11 @@ export const getStaticProps = async ({ locale }: { locale: string }) => ({
 })
 
 export default function LandingPage() {
-  const enabled = useFeatureToggle('AssetLandingPages')
-  const view = enabled ? <HomepageView /> : <LandingView />
-  return <WithConnection>{view}</WithConnection>
+  return (
+    <WithConnection>
+      <HomepageView />
+    </WithConnection>
+  )
 }
 
 LandingPage.layout = LandingPageLayout
